Backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
Authlib==1.3.2
blinker==1.9.0
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
groq==0.25.0
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.56
langchain-text-splitters==0.3.8
langsmith==0.1.145
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==2.1.0
orjson==3.10.11
packaging==24.2
pillow==11.0.0
portalocker==2.10.1
propcache==0.2.0
protobuf==6.30.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.1
pydantic-settings==2.9.1
pydantic_core==2.27.1
PyMuPDF==1.24.14
PyPDF2==3.0.1
python-dotenv==1.0.1
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.1.0
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.0
